+++ aggregation/input/002/input.json
