+++ aggregation/package.json
