--- circuits/aggregation/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2021",
-    "module": "commonjs",
-    "allowJs": true,
-    "esModuleInterop": true,
-    "esModuleInterop": true,
-    "outDir": "dist"
-  },
-  "include": ["./scripts", "./test", "typechain/**/*.ts", "typechain/**/*.d.ts", "typechain/**/*.js", "lib"],
-  "files": ["./hardhat.config.ts"]
-}
